audit2db-plugin/src/main/java/org/jenkins/plugins/audit2db/pipeline/Audit2DbStep.java
package org.jenkins.plugins.audit2db.pipeline;

import java.util.Set;

import org.jenkinsci.plugins.workflow.steps.Step;
import org.jenkinsci.plugins.workflow.steps.StepContext;
import org.jenkinsci.plugins.workflow.steps.StepDescriptor;
import org.jenkinsci.plugins.workflow.steps.StepExecution;
import org.kohsuke.stapler.DataBoundConstructor;
import org.kohsuke.stapler.DataBoundSetter;

import com.google.common.collect.ImmutableSet;

import hudson.Extension;
import hudson.model.Computer;
import hudson.model.Run;

public class Audit2DbStep extends Step {

	private String jdbcDriver;
	private String jdbcUrl;
	private String jdbcUsername;
	private String jdbcPassword;
	private String hibernateDialect;

	public Audit2DbStep() {
		super();
	}

	@DataBoundConstructor
	public Audit2DbStep(String jdbcDriver, String jdbcUrl, String jdbcUsername,
			String jdbcPassword) {
		super();
		this.jdbcDriver = jdbcDriver;
		this.jdbcUrl = jdbcUrl;
		this.jdbcUsername = jdbcUsername;
		this.jdbcPassword = jdbcPassword;
	}

	@Override
	public StepExecution start(StepContext context) throws Exception {
		Audit2DbExecution execution = new Audit2DbExecution(this, context);
		return execution;
	}

	public String getJdbcDriver() {
		return jdbcDriver;
	}

	public void setJdbcDriver(String jdbcDriver) {
		this.jdbcDriver = jdbcDriver;
	}

	public String getJdbcUrl() {
		return jdbcUrl;
	}

	public void setJdbcUrl(String jdbcUrl) {
		this.jdbcUrl = jdbcUrl;
	}

	public String getJdbcUsername() {
		return jdbcUsername;
	}

	public void setJdbcUsername(String jdbcUsername) {
		this.jdbcUsername = jdbcUsername;
	}

	public String getJdbcPassword() {
		return jdbcPassword;
	}

	public void setJdbcPassword(String jdbcPassword) {
		this.jdbcPassword = jdbcPassword;
	}

	public String getHibernateDialect() {
		return hibernateDialect;
	}

	@DataBoundSetter
	public void setHibernateDialect(String hibernateDialect) {
		this.hibernateDialect = hibernateDialect;
	}

	@Extension
	public static class DescriptorImpl extends StepDescriptor {

		@Override
		public Set<? extends Class<?>> getRequiredContext() {
			return ImmutableSet.of(Run.class, Computer.class);
		}

		@Override
		public String getFunctionName() {
			return "audit2db";
		}

		@Override
		public String getDisplayName() {
			return "Audit to Database";
		}

	}

}
